+++ Reginald.Core/Enums/SpecialSearchResultEnums.cs
@@ -2,6 +2,7 @@
 {
     public enum Api
     {
+        None = 0,
         Styvio,
         Cloudflare
     }