+++ Reginald.Core/IO/ApplicationPaths.cs
@@ -14,6 +14,7 @@ namespace Reginald.Core.IO
         public const string XmlCommandsFilename = "Commands.xml";
         public const string XmlSettingsFilename = "Settings.xml";
         public const string XmlUtilitiesFilename = "Utilities.xml";
+        public const string XmlSettingsPagesFileLocation = "Resources/MSSettings.xml";
         public const string TxtFilename = "Applications.txt";
     }
 }