+++ Reginald/ViewModels/CommandsViewModel.cs
@@ -9,61 +9,19 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Input;
 using System.Xml;
 
 namespace Reginald.ViewModels
 {
-    public class CommandsViewModel : Screen
+    public class CommandsViewModel : KeywordViewModel
     {
-        public CommandsViewModel()
+        public CommandsViewModel() : base(ApplicationPaths.XmlCommandsFilename, true)
         {
-            LoadKeywordSearchResultsAsync(ApplicationPaths.XmlCommandsFilename, Commands);
-            Settings.IncludeCommands = Properties.Settings.Default.IncludeDefaultKeywords;
+            LoadKeywordSearchResults(ApplicationPaths.XmlCommandsFilename, KeywordSearchResults);
+            Settings.IncludeCommands = Properties.Settings.Default.IncludeCommands;
         }
 
-        private SettingsModel _settings = new();
-        public SettingsModel Settings
-        {
-            get => _settings;
-            set
-            {
-                _settings = value;
-                NotifyOfPropertyChange(() => Settings);
-            }
-        }
-
-        private BindableCollection<SearchResultModel> _commands = new();
-        public BindableCollection<SearchResultModel> Commands
-        {
-            get => _commands;
-            set
-            {
-                _commands = value;
-                NotifyOfPropertyChange(() => Commands);
-            }
-        }
-
-        private SearchResultModel _selectedCommand = new();
-        public SearchResultModel SelectedCommand
-        {
-            get => _selectedCommand;
-            set
-            {
-                _selectedCommand = value;
-                NotifyOfPropertyChange(() => SelectedCommand);
-            }
-        }
-
-        public void ScrollViewer_PreviewMouseWheel(object sender, MouseWheelEventArgs e)
-        {
-            ScrollViewer scv = (ScrollViewer)sender;
-            scv.ScrollToVerticalOffset(scv.VerticalOffset - e.Delta);
-            e.Handled = true;
-        }
-
-        public void IncludeCommandsToggleSwitch_Checked(object sender, RoutedEventArgs e)
+        public void IncludeCommandsToggleButton_Checked(object sender, RoutedEventArgs e)
         {
             bool value = !Properties.Settings.Default.IncludeCommands;
             Properties.Settings.Default.IncludeCommands = value;
@@ -71,26 +29,11 @@ namespace Reginald.ViewModels
             Settings.IncludeCommands = value;
         }
 
-        public void Commands_IsCheckedChanged(object sender, RoutedEventArgs e)
-        {
-            ChangeIsEnabledState(ApplicationPaths.XmlCommandsFilename, SelectedCommand.ID);
-        }
-
-        private static void ChangeIsEnabledState(string name, int id)
+        public override async void LoadKeywordSearchResults(string name, BindableCollection<SearchResultModel> models)
         {
             XmlDocument doc = XmlHelper.GetXmlDocument(name);
-            string xpath = string.Format(Constants.NamespaceIDXpathFormat, id);
-            XmlNode node = doc.SelectSingleNode(xpath);
-            XmlNode isEnabledNode = node.SelectSingleNode("IsEnabled");
-            bool isEnabled = bool.Parse(isEnabledNode.InnerText);
-            isEnabledNode.InnerText = (!isEnabled).ToString().ToLower();
-            XmlHelper.SaveXmlDocument(doc, name);
-        }
-
-        private static async void LoadKeywordSearchResultsAsync(string name, BindableCollection<SearchResultModel> collection)
-        {
-            XmlDocument doc = XmlHelper.GetXmlDocument(name);
-            IEnumerable<string> attributes = doc.GetNodesAttributes(Constants.NamespacesXpath);
+            IEnumerable<string> attributes = doc.GetNodesAttributes(Constants.NamespacesXpath)
+                                                .Distinct();
 
             IEnumerable<SearchResultModel> commands = Array.Empty<SearchResultModel>();
             string description = string.Empty;
@@ -98,7 +41,7 @@ namespace Reginald.ViewModels
             {
                 commands = commands.Concat(await SearchResultModel.MakeListForCommandsAsync(doc, description, attribute, Category.Notifier));
             }
-            collection.AddRange(commands);
+            models.AddRange(commands);
         }
     }
 }