+++ Reginald/ViewModels/SpecialKeywordViewModel.cs
@@ -11,66 +11,66 @@ using System.Xml;
 
 namespace Reginald.ViewModels
 {
-    public class SpecialKeywordViewModel : Screen
+    public class SpecialKeywordViewModel : KeywordViewModel
     {
-        public SpecialKeywordViewModel()
+        public SpecialKeywordViewModel() : base(ApplicationPaths.XmlSpecialKeywordFilename)
         {
-            LoadSpecialKeywordSearchResults(ApplicationPaths.XmlSpecialKeywordFilename, SpecialKeywordSearchResults);
+            //LoadSpecialKeywordSearchResults(ApplicationPaths.XmlSpecialKeywordFilename, SpecialKeywordSearchResults);
             Settings.IncludeSpecialKeywords = Properties.Settings.Default.IncludeSpecialKeywords;
         }
 
-        private SettingsModel _settings = new();
-        public SettingsModel Settings
-        {
-            get => _settings;
-            set
-            {
-                _settings = value;
-                NotifyOfPropertyChange(() => Settings);
-            }
-        }
+        //private SettingsModel _settings = new();
+        //public SettingsModel Settings
+        //{
+        //    get => _settings;
+        //    set
+        //    {
+        //        _settings = value;
+        //        NotifyOfPropertyChange(() => Settings);
+        //    }
+        //}
 
-        private BindableCollection<SpecialSearchResultModel> _specialKeywordSearchResults = new();
-        public BindableCollection<SpecialSearchResultModel> SpecialKeywordSearchResults
-        {
-            get => _specialKeywordSearchResults;
-            set
-            {
-                _specialKeywordSearchResults = value;
-                NotifyOfPropertyChange(() => SpecialKeywordSearchResults);
-            }
-        }
+        //private BindableCollection<SpecialSearchResultModel> _specialKeywordSearchResults = new();
+        //public BindableCollection<SpecialSearchResultModel> SpecialKeywordSearchResults
+        //{
+        //    get => _specialKeywordSearchResults;
+        //    set
+        //    {
+        //        _specialKeywordSearchResults = value;
+        //        NotifyOfPropertyChange(() => SpecialKeywordSearchResults);
+        //    }
+        //}
 
-        private SpecialSearchResultModel _specialKeywordSearchResult = new();
-        public SpecialSearchResultModel SpecialKeywordSearchResult
-        {
-            get => _specialKeywordSearchResult;
-            set
-            {
-                _specialKeywordSearchResult = value;
-                NotifyOfPropertyChange(() => SpecialKeywordSearchResult);
-            }
-        }
+        //private SpecialSearchResultModel _specialKeywordSearchResult = new();
+        //public SpecialSearchResultModel SpecialKeywordSearchResult
+        //{
+        //    get => _specialKeywordSearchResult;
+        //    set
+        //    {
+        //        _specialKeywordSearchResult = value;
+        //        NotifyOfPropertyChange(() => SpecialKeywordSearchResult);
+        //    }
+        //}
 
-        private SpecialSearchResultModel _selectedKeywordSpecialSearchResult = new();
-        public SpecialSearchResultModel SelectedSpecialKeywordSearchResult
-        {
-            get => _selectedKeywordSpecialSearchResult;
-            set
-            {
-                _selectedKeywordSpecialSearchResult = value;
-                NotifyOfPropertyChange(() => SelectedSpecialKeywordSearchResult);
-            }
-        }
+        //private SpecialSearchResultModel _selectedKeywordSpecialSearchResult = new();
+        //public SpecialSearchResultModel SelectedSpecialKeywordSearchResult
+        //{
+        //    get => _selectedKeywordSpecialSearchResult;
+        //    set
+        //    {
+        //        _selectedKeywordSpecialSearchResult = value;
+        //        NotifyOfPropertyChange(() => SelectedSpecialKeywordSearchResult);
+        //    }
+        //}
 
-        public void ScrollViewer_PreviewMouseWheel(object sender, MouseWheelEventArgs e)
-        {
-            ScrollViewer scv = (ScrollViewer)sender;
-            scv.ScrollToVerticalOffset(scv.VerticalOffset - e.Delta);
-            e.Handled = true;
-        }
+        //public void ScrollViewer_PreviewMouseWheel(object sender, MouseWheelEventArgs e)
+        //{
+        //    ScrollViewer scv = (ScrollViewer)sender;
+        //    scv.ScrollToVerticalOffset(scv.VerticalOffset - e.Delta);
+        //    e.Handled = true;
+        //}
 
-        public void IncludeSpecialKeywordsToggleSwitch_Checked(object sender, RoutedEventArgs e)
+        public void IncludeSpecialKeywordsToggleButton_Checked(object sender, RoutedEventArgs e)
         {
             bool value = !Properties.Settings.Default.IncludeSpecialKeywords;
             Properties.Settings.Default.IncludeSpecialKeywords = value;
@@ -78,31 +78,31 @@ namespace Reginald.ViewModels
             Settings.IncludeSpecialKeywords = value;
         }
 
-        public void SpecialKeywordSearchResults_IsCheckedChanged(object sender, RoutedEventArgs e)
-        {
-            ChangeIsEnabledState(ApplicationPaths.XmlSpecialKeywordFilename, SelectedSpecialKeywordSearchResult.ID);
-        }
+        //public void SpecialKeywordSearchResults_IsCheckedChanged(object sender, RoutedEventArgs e)
+        //{
+        //    ChangeIsEnabledState(ApplicationPaths.XmlSpecialKeywordFilename, SelectedSpecialKeywordSearchResult.ID);
+        //}
 
-        private static void ChangeIsEnabledState(string name, int id)
-        {
-            XmlDocument doc = XmlHelper.GetXmlDocument(name);
-            string xpath = string.Format(Constants.NamespaceIDXpathFormat, id);
-            XmlNode node = doc.SelectSingleNode(xpath);
-            XmlNode isEnabledNode = node.SelectSingleNode("IsEnabled");
-            bool isEnabled = bool.Parse(isEnabledNode.InnerText);
-            isEnabledNode.InnerText = (!isEnabled).ToString().ToLower();
-            XmlHelper.SaveXmlDocument(doc, name);
-        }
+        //private static void ChangeIsEnabledState(string name, int id)
+        //{
+        //    XmlDocument doc = XmlHelper.GetXmlDocument(name);
+        //    string xpath = string.Format(Constants.NamespaceIDXpathFormat, id);
+        //    XmlNode node = doc.SelectSingleNode(xpath);
+        //    XmlNode isEnabledNode = node.SelectSingleNode("IsEnabled");
+        //    bool isEnabled = bool.Parse(isEnabledNode.InnerText);
+        //    isEnabledNode.InnerText = (!isEnabled).ToString().ToLower();
+        //    XmlHelper.SaveXmlDocument(doc, name);
+        //}
 
-        private static void LoadSpecialKeywordSearchResults(string name, BindableCollection<SpecialSearchResultModel> collection)
-        {
-            XmlDocument doc = XmlHelper.GetXmlDocument(name);
-            XmlNodeList nodes = doc.GetNodes(Constants.NamespacesXpath);
+        //private static void LoadSpecialKeywordSearchResults(string name, BindableCollection<SpecialSearchResultModel> collection)
+        //{
+        //    XmlDocument doc = XmlHelper.GetXmlDocument(name);
+        //    XmlNodeList nodes = doc.GetNodes(Constants.NamespacesXpath);
 
-            foreach (XmlNode node in nodes)
-            {
-                collection.Add(new SpecialSearchResultModel(node, string.Empty));
-            }
-        }
+        //    foreach (XmlNode node in nodes)
+        //    {
+        //        collection.Add(new SpecialSearchResultModel(node, string.Empty));
+        //    }
+        //}
     }
 }